third_party/WebKit/LayoutTests/fast/dom/global-event-handlers.html

<!DOCTYPE html>
<title>GlobalEventHandlers test</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../html/resources/event-handlers.js"></script>
<script>
// attribute list from WHATWG HTML Living Standard r8389
[
    "onabort",
    "onblur",
    "oncancel",
    "oncanplay",
    "oncanplaythrough",
    "onchange",
    "onclick",
    "onclose",
    "oncontextmenu",
    "oncuechange",
    "ondblclick",
    "ondrag",
    "ondragend",
    "ondragenter",
    "ondragexit",
    "ondragleave",
    "ondragover",
    "ondragstart",
    "ondrop",
    "ondurationchange",
    "onemptied",
    "onended",
    "onerror",
    "onfocus",
    "oninput",
    "oninvalid",
    "onkeydown",
    "onkeypress",
    "onkeyup",
    "onload",
    "onloadeddata",
    "onloadedmetadata",
    "onloadstart",
    "onmousedown",
    "onmouseenter",
    "onmouseleave",
    "onmousemove",
    "onmouseout",
    "onmouseover",
    "onmouseup",
    "onmousewheel",
    "onpause",
    "onplay",
    "onplaying",
    "onprogress",
    "onratechange",
    "onreset",
    "onresize",
    "onscroll",
    "onseeked",
    "onseeking",
    "onselect",
    "onsort",
    "onstalled",
    "onsubmit",
    "onsuspend",
    "ontimeupdate",
    "ontoggle",
    "ontouchcancel",
    "ontouchend",
    "ontouchmove",
    "ontouchstart",
    "onvolumechange",
    "onwaiting",
    "onwheel"
].forEach(function(attribute) {
    test(function() {
        assert_false(attribute in getObject("Element"));
    }, "No Element." + attribute);
    testSet("HTMLElement", attribute);
    testEnumerate("HTMLElement", attribute);
    testReflect("HTMLElement", attribute);
    testSet("SVGElement", attribute);
    testEnumerate("SVGElement", attribute);
    testReflect("SVGElement", attribute);
    testSet("Document", attribute);
    testEnumerate("Document", attribute);
    testSet("Window", attribute);
    testEnumerate("Window", attribute);
});
</script>
<div id="log"></div>
